Macro better_any::type_id

macro_rules! type_id {
    ($struct: ident) => { ... };
    ($struct: ident < $lt: lifetime >) => { ... };
}

Simple version of derive macro to not pull all proc macro dependencies in simple cases when all structs are not generic

use better_any::type_id;
struct S;
type_id!(S);
struct F<'a>(&'a str);
type_id!(F<'a>);