Trait better_any::TidExt

pub trait TidExt<'a> {
    fn is<T: Tid<'a>>(&self) -> bool;
    fn downcast_ref<'b, T: Tid<'a>>(&'b self) -> Option<&'b T>;
    fn downcast_mut<'b, T: Tid<'a>>(&'b mut self) -> Option<&'b mut T>;
    fn downcast_rc<T: Tid<'a>>(self: Rc<Self>) -> Result<Rc<T>, Rc<Self>>;
    fn downcast_arc<T: Tid<'a>>(self: Arc<Self>) -> Result<Arc<T>, Arc<Self>>;
    fn downcast_box<T: Tid<'a>>(self: Box<Self>) -> Result<Box<T>, Box<Self>>;
}

Extension trait that contains actual downcasting methods.

Use methods from this trait only if dyn Tid was created directly from T for this particular T

Required Methods

fn is<T: Tid<'a>>(&self) -> bool

Returns true if type behind self is equal to the type of T.

fn downcast_ref<'b, T: Tid<'a>>(&'b self) -> Option<&'b T>

Attempts to downcast self to T behind reference

fn downcast_mut<'b, T: Tid<'a>>(&'b mut self) -> Option<&'b mut T>

Attempts to downcast self to T behind mutable reference

fn downcast_rc<T: Tid<'a>>(self: Rc<Self>) -> Result<Rc<T>, Rc<Self>>

Attempts to downcast self to T behind Rc pointer

fn downcast_arc<T: Tid<'a>>(self: Arc<Self>) -> Result<Arc<T>, Arc<Self>>

Attempts to downcast self to T behind Arc pointer

fn downcast_box<T: Tid<'a>>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Attempts to downcast self to T behind Box pointer